pyproject.toml
[project]
name = "aas"
dynamic = ["version"]
description = "A helpful bot with speech input and output"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name="Jan Scholz", email="jscholz.contact@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.44.0",
    "fastapi>=0.115.13",
    "openai>=1.90.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]

[tool.hatch.version]
path = "your_package/__init__.py"

[tool.hatch.envs.dev]
dependencies = [
    "sounddevice>=0.5.2",
    "numpy>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
